Tweak config template to omit settings without a value
Without this change, settings that don't provide a default value are written
to the configuration file as 'setting = "undef"'.  This is a problem because
vault needs them to be omitted in order to use its defaults.

Also includes a documentation update to indicate which backends have been
tested, as test coverage on them is currently fairly limited.