Add --check option to vault-auth-app helper script
Without this change, it's difficult to use the vault-auth-app script in an
exec resource without triggering it on every run.  This change provides a
simple solution: just add an "unless" argument to the exec, with the same
value as the "command" argument, but with the --check flag added to it.