Get the empty string from jq for empty values
Without this change, jq returns the string "null" when there's no data
available.  This is a problem because we treat the output as a string and
end up adding a role called "null" where we shouldn't be adding one.  This
change forces jq to return an empty string rather than the word "null".