Ensure vault is ready before attempting to use it
Without this change, the vault-bootstrap may attempt to write data to vault
before it's ready to have data written to it.  This is a problem because the
data that's being written to it first is critical to the overall setup, and
not having it available causes the PKI backend (and potentially other parts)
to be completely unusable.