Gracefully handle machine UUIDs that are "Not Settable"
On some machines, the output of "facter uuid" is "Not Settable".  This is
not appropriate for use as a unique machine identifier because (a) it
contains a space, breaking some of the vault commands, and (b) it is not
likely to be unique (all VMs on a given physical host are likely to have the
same issue).