Ensure connections to Vault are direct
Without this change, the "curl" and "vault" commands will use the proxy
configured in the environment ($http_proxy and/or $https_proxy) to access
Vault.  This is a problem when Vault is being accessed at 127.0.0.1 (which
is expected to be the case during setup), as the proxy won't be able to
reach Vault on that address.

This change makes the assumption that direct access to Vault is possible
(which currently seems quite reasonable).