Give Consul a chance to complete startup before refusing to bootstrap
Without this change, the vault-bootstrap may refuse to run on account of
Consul having started but not initialised.  This is a problem because it
introduces a fairly common situation where puppet doesn't complete properly
on the first run (and we want to be able to configure servers with a single
puppet run).